requirements.txt
PyYAML==6.0.2
pandas==2.2.3
openpyxl==3.1.5
pytest==8.3.4
tabulate==0.9.0